Remove a non-exhaustive case from passC.mkanno
Non-exhaustive case analysis was done on a parameter 'msym :: Symbol'
but that was done just before returning a value.
So it was moved out of the function to the caller site, removing the
need to match on Symbol.

Also the parameter was changed to 'mpos :: Position' because 'mkanno'
was only interested in 'pos' field.