Change the type of SymI.env to SymMeth
SymMeth is a new data type which is a sum of SymL and SymV.
SymI.env was renamed to SymI.meth and can now accept SymL and SymV only.

The following functions now take SymMeth instead of Symbol. This change
is considered safe because they threw runtime errors when receiving
other than SymL or SymV.

- frege.compiler.Classes.tcInstMethod
- frege.compiler.gen.java.VarCode.varCode
- frege.compiler.passes.Easy.checkDepth

SymbolT.env' which is a generalization of .env and .meth is added and
uses of SymbolT.env are replaced with SymbolT.env' where applicable.