Remove Prisms and use Traversals instead
We don't use 'review', so Traversal is enough.
Well, we do use 'is', but 'has' can be used for the purpose.

Prism depends on Profunctor. In Haskell, the "profunctor" is a huge
package. If, in some day, we port it to Frege, the compiler depending on
it may be a problem.