Change the grammar 'gargs' to optionally take [TVar] instead of [TauS]
'gargs', generics arguments that optionally put on native declarations
like in 'datajavainit'

    data Foo = native java.util.HashMap {k, v}

'{k, v}' are 'gargs'. The elements 'k, v' used to be 'tauSC' but changed
to a new grammar 'gargvars'. The consequence is that non-TVar tau types
are no longer accepted; for example, the following is now an error in
the parse pass:

    {[k], m v}

but it should have been an error for the first place; in fact, JavDcl
(indirectly) required that to be a case. Otherwise, the compiler crashed
by accessing a partial record field (TauT.var, a the master branch).

Actually, NatDcl used to accept non-TVar gargs. The following compiled:

    data List a = native java.util.List
    native emptyListFoo java.util.Collections.emptyList {[a]} :: () -> STMutable s (List [a])

but it will be rejected from now on as a syntax error.