Add symItau to remove the partial case in instFun and docSym
All of ide.Utilities, tools.doc.Utilities, and
compiler.gen.java.InstanceCode rely on the same (type-unsafe) invariant,
so it was extracted as a new function.

Note that the function is intended to be a a temporary measure until I
implement MultiParamTypeClasses and perform a required refactoring in
which SymI.typ will be broken down into ctxs :: [Context] and
tau :: Tau.